Generate ggplot friendly data frame for plotting functional responses

Generate ggplot friendly data frame for plotting functional responses

df_funcresp(pars, maxx)

Arguments

pars
parameter list from spec_rescomp()

maxx
maximum resource value to get percapita growth rates across (for plotting)

Value

data frame

Examples

pars <- spec_rescomp()

#> Model properties 
#>  * 1 consumer(s) and 1 resource(s)
#>  * Consumers have type 1 functional responses
#>  * Resource supply is continuous (e.g. chemostat)
#>  * Mortality is continuous
#> 
#> Simulation properties 
#>  * Simulation time: 1000 time steps
#>  * Init state: consumer(s) = [10], resource(s) = [1]

df_funcresp(pars)

#> # A tibble: 1,000 × 5
#>    resource.levels sp    paramstate  resource     growth
#>              <dbl> <chr> <chr>       <chr>         <dbl>
#>  1         0       N1    env-state-1 Resource a 0       
#>  2         0.00100 N1    env-state-1 Resource a 0.000100
#>  3         0.00200 N1    env-state-1 Resource a 0.000200
#>  4         0.00300 N1    env-state-1 Resource a 0.000300
#>  5         0.00400 N1    env-state-1 Resource a 0.000400
#>  6         0.00501 N1    env-state-1 Resource a 0.000501
#>  7         0.00601 N1    env-state-1 Resource a 0.000601
#>  8         0.00701 N1    env-state-1 Resource a 0.000701
#>  9         0.00801 N1    env-state-1 Resource a 0.000801
#> 10         0.00901 N1    env-state-1 Resource a 0.000901
#> # … with 990 more rows