Generate list of parameters for a consumer-resource model to be passed to sim_rescomp()

Generate list of parameters for a consumer-resource model to be passed to sim_rescomp()

spec_rescomp(
  spnum = 1,
  resnum = 1,
  funcresp = funcresp_type1(crmatrix(0.1)),
  quota = crmatrix(0.001),
  efficiency,
  essential = FALSE,
  mort = 0.03,
  ressupply = ressupply_chemostat(0.03, 1),
  params = rescomp_param_list(),
  events = list(),
  totaltime = 1000,
  cinit = 10,
  rinit = 1,
  verbose = FALSE
)

Arguments

spnum

Integer vector of length 1: the number of consumers.

resnum

Integer vector of length 1: the number of resources.

funcresp

An object of class rescomp_funcresp specifying the functional response.

quota

Numeric matrix or rescomp_coefs_matrix, the elements of which give the resource quotas. The number of rows and columns should be equal to spnum and resnum respectively. Mutually exclusive with efficiency.

efficiency

Numeric matrix or rescomp_coefs_matrix, the elements of which give the efficiency of each consumer on each resource. The number of rows and columns should be equal to spnum and resnum respectively. Mutually exclusive with quota.

essential

Logical vector of length 1. If FALSE resources are substitutable.

mort

Numeric vector or rescomp_coefs_vector of length spnum, specifying density independent mortality rates.

ressupply

An object of class rescomp_ressupply specifying the resource supply.

params

An object of class rescomp_param_list specifying a set of parameters which vary with time, on which other parameters of the model (e.g. funcresp, ressupply) may depend.

events

A list of objects of class rescomp_event_schedule, specifying events that instantaneously change consumer or resource densities.

totaltime

Numeric vector of length 1: the total simulation time.

cinit

Numeric vector of length 1 or length spnum specifying initial consumer state values (densities).

rinit

Numeric vector of length 1 or length resnum specifying initial resource state values (concentrations).

verbose

If TRUE (default) prints model and simulation summary specs.

Value

S3 object of class rescomp.

Details

Only one of efficiency and quota should be specified. Specifying both is an error. The default, if neither is specified, is to use quota. If using quota, the functional responses are taken to give per capita growth rates. If using efficiency, the functional responses are taken to give attack rates.

Examples

# TODO