Generate list of parameters for a consumer-resource model to be passed to sim_rescomp

Generate list of parameters for a consumer-resource model to be passed to sim_rescomp

spec_rescomp(
  spnum = 1,
  resnum = 1,
  mumatrix,
  kmatrix,
  qmatrix,
  funcresp = "type1",
  essential = FALSE,
  chemo = TRUE,
  mort = 0.03,
  resspeed = 0.03,
  resconc = 1,
  respulse = 0,
  mortpulse = 0,
  pulsefreq = 0,
  batchtrans = FALSE,
  timepars = FALSE,
  totaltime = 1000,
  timeparfreq = 0,
  tpinterp = "inst",
  cinit = 10,
  verbose = TRUE
)

Arguments

spnum
Number of consumers

resnum
Number of resources

mumatrix
A list containing the matrix(s) of maximum growth rates (type 2 or 3) / constant of proportionalities (type 1). The number of rows should be equal to spnum and resnum respectively. If timepars = TRUE, expects a list of length 2.

kmatrix
Matrix of half saturation constants (type 2 or type 3, ignored if funcresp = "type1). The number of rows should be equal to spnum and resnum respectively.

qmatrix
Matrix of resource quotas. The number of rows should be equal to spnum and resnum respectively.

funcresp
Options include "type1", 'type2", or "type3".

essential
If FALSE resources are substitutable.

chemo
Default is resources supplied continuously (chemostat). If FALSE resources grow logistically.

mort
Density independent mortality rate.

resspeed
Resource intrinsic rate of increase (if chemo = FALSE), otherwise chemostat dilution rate. Set to zero for pulsing only.

resconc
Resource carrying capacity (if chemo = FALSE), otherwise chemostat supply concentration

respulse
Resource pulse size.

mortpulse
Consumer mortality fraction if non-constant mortality (e.g. serial-batch transfer).

pulsefreq
Frequency of resource pulsing and/or intemittent mortality.

batchtrans
If TRUE, both resource is also fractionally sampled (see vignette).

timepars
If TRUE, time dependent parameters required.

totaltime
Total simulation time

timeparfreq
Frequency of parameter switching if timepars = TRUE.

tpinterp
Either "inst" (default), "lin" or "sine" interpolation of time dependent parameters. If "inst", parameters switch instantaneously with frequency given by timeparfreq. If "lin" or "sine", parameters are interpolated linearly or sinusoidaly, respectively, with period 2x timeparfreq.

cinit
Initial consumer state values (densities). Either a single integer for all consumers or a vector. Defaults to 10 for all consumers. Note initial resource state values defaults to resconc.

verbose
If TRUE (default) prints model and simulation summary specs.

Value

List object of class 'rescomp'.

Examples

# Single type I consumer on a single logistically growing resource
spec_rescomp()

#> Model properties 
#>  * 1 consumer(s) and 1 resource(s)
#>  * Consumers have type 1 functional responses
#>  * Resource supply is continuous (e.g. chemostat)
#>  * Mortality is continuous
#> 
#> Simulation properties 
#>  * Simulation time: 1000 time steps
#>  * Init state: consumer(s) = [10], resource(s) = [1]

# Two type II consumers and two substitutable resources in a chemostat
pars <- spec_rescomp(
    spnum = 2,
    resnum = 2,
    funcresp = "type2",
    mumatrix = list(matrix(c(0.7,0.3,
                             0.4,0.5),
                           nrow = 2,
                           ncol = 2,
                           byrow = TRUE)),
    resspeed = 3,
    resconc = 1,
    chemo = TRUE,
    essential = FALSE
)

#> Model properties 
#>  * 2 consumer(s) and 2 resource(s)
#>  * Consumers have type 2 functional responses
#>  * Resources are substitutable
#>  * Resource supply is continuous (e.g. chemostat)
#>  * Mortality is continuous
#> 
#> Simulation properties 
#>  * Simulation time: 1000 time steps
#>  * Init state: consumer(s) = [10, 10], resource(s) = [1, 1]